vectorRestore only when necessary
Added dependency on AlyBox, a small library mod that implements **player.entity.name**, a variable used to determine the player's current character config.

The **vectorRestore** EventTrigger in the arena lobby is disruptive to other mods that allow you to play as other characters, so with this fix it will only trigger if *8player.entity.name == 'Vector'**.

I have tested this by walking into the arena as the Vector player config as well as any other player config and confirmed that it works.

AlyBox: https://github.com/lexisother/cc-alybox